% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{backtransform_marginals}
\alias{backtransform_marginals}
\title{Backtransform copula samples using fitted marginals}
\usage{
backtransform_marginals(u, marginals)
}
\arguments{
\item{u}{Numeric matrix of size n x p with copula samples}

\item{marginals}{List of marginals from \code{\link[=estimateMarginals]{estimateMarginals()}}}
}
\value{
data.frame with n rows and length(marginals) columns
}
\description{
Takes a matrix of copula samples (on \[0,1\]) and a list of marginal models
and returns a data.frame on the original marginal scales.
}
\keyword{internal}
