% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generateSynthetic}
\alias{generateSynthetic}
\title{Generate synthetic data}
\usage{
generateSynthetic(
  n,
  copula,
  marginals,
  original_levels = NULL,
  original_varnames = NULL,
  original_classes = NULL
)
}
\arguments{
\item{n}{Integer, number of rows to generate}

\item{copula}{A PUcopula model}

\item{marginals}{List of marginals from \code{\link[=estimateMarginals]{estimateMarginals()}}}

\item{original_levels}{Optional \code{preprocessData()$original_levels}}

\item{original_varnames}{Optional vector of original column names}

\item{original_classes}{Optional named vector of original classes}
}
\value{
Synthetic data.frame
}
\description{
Combines a fitted PUcopula and marginal models to produce a synthetic
data.frame. Optionally restores original factor structure, names and classes.
}
