% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knnsmoother}
\alias{knnsmoother}
\title{k-NN smoother for numeric vectors (no DataSHIELD thresholds)}
\usage{
knnsmoother(x, k = 3)
}
\arguments{
\item{x}{Numeric vector (NAs allowed).}

\item{k}{Integer neighbours in \link{1, N-1} for non-missing values.}
}
\value{
Numeric vector of same length with smoothed non-missing entries.
}
\description{
k-NN smoother for numeric vectors (no DataSHIELD thresholds)
}
