% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{save_original_varnames}
\alias{save_original_varnames}
\alias{save_original_classes}
\title{Save original names and classes}
\usage{
save_original_varnames(data)

save_original_classes(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Character vector (names) / Named character vector (classes)
}
\description{
Save original names and classes
}
