% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{simulateCopula}
\alias{simulateCopula}
\title{Draw from a fitted PUcopula}
\usage{
simulateCopula(model, n)
}
\arguments{
\item{model}{A PUcopula model from \code{\link[=fitPUcopula]{fitPUcopula()}}}

\item{n}{Number of rows}
}
\value{
Matrix of U(0,1) draws
}
\description{
Draw from a fitted PUcopula
}
